Added conontroller concerns for devise params
