+++ app/controllers/concerns/devise_whitelist.rb
@@ -1,0 +1,11 @@
+module DeviseWhitelist
+  extend ActiveSupport::Concern
+  included do
+    before_action :configure_permitted_parameters,
+                  if: :devise_controller?
+  end
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:name])
+    devise_parameter_sanitizer.permit(:account_update, keys: [:name])
+  end
+end
